+++ src/java/facade/SolutionFacade.java
@@ -6,10 +6,12 @@
 package facade;
 
 import controller.util.JsfUtil;
+import java.util.List;
 import javax.ejb.Stateless;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import javax.persistence.PersistenceException;
+import model.Reponse;
 import model.Solution;
 
 /**
@@ -32,7 +34,7 @@ public class SolutionFacade extends AbstractFacade<Solution> {
     }
 
     public boolean create2(Solution solution) {
-              int id = 1;
+        int id = 1;
         try {
             Solution q = getEntityManager().createNamedQuery("Solution.findMaxId", Solution.class).getSingleResult();
             id = q.getId() + 1;
@@ -57,8 +59,18 @@ public class SolutionFacade extends AbstractFacade<Solution> {
     }
 
     public boolean remove2(Solution solution) {
-           getEntityManager().remove(getEntityManager().merge(solution));
-         return true;
+        List<Reponse> list;
+        list = getEntityManager().createNamedQuery("Reponse.findBySolutionID", Reponse.class).
+                setParameter("solution_id", solution.getId()).getResultList();
+        if (!list.isEmpty()) {
+            //  System.out.println("koko" + list.get(0).getId());
+            JsfUtil.addErrorMessage("Cette solution ne peut pas étre supprimée car elle est toujours réferencée par une réponse.");
+            return false;
+        } else {
+            System.out.println("list == null remove ");
+            getEntityManager().remove(getEntityManager().merge(solution));
+            return true;
+        }
     }
-    
+
 }
